tests/requirements.txt
pytest==7.4.4
pyyaml==6.0.1
kubernetes>=33.0.0
docker==6.1.3
jsonschema==4.20.0
requests==2.31.0
git+https://github.com/syntasso/kratix-python.git